Update second section and handle thin screens
For thin screens, the CSS code is updated to make the main text
fill more of the width of the screen. This may help with portrait
mode on smartphones where there is limited screen width within
which to place text at all, so the blank margin on the right looks
awkwardly large with the default margins.