Clarify return value optimization & eqn sizes
This sets the max-width property on the Navier-Stokes and Euler
equations to make them scale with the screen size and make them
readable on narrow screens, while not making them comically huge
on wider (e.g., desktop or laptop monitor) screens.

This also clarifies the use and effects of return value
optimization, including changes made to the SimulationParameters
class code to ensure return value optimization is being applied.